% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc_rej.R
\name{abc_rej}
\alias{abc_rej}
\title{Rejection ABC}
\usage{
abc_rej(prior, s, s_obs, q = 0.01)
}
\arguments{
\item{prior}{Matrix of parameters from the prior}

\item{s}{Matrix of corresponding summary statistics}

\item{q}{Distance quantile under which simulations are excepted}

\item{obs}{Vector of observed summary statistics}
}
\description{
Rejection ABC
}
