% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mean_and_sd}
\alias{mean_and_sd}
\title{Calculate colmeans and colstds}
\usage{
mean_and_sd(s)
}
\arguments{
\item{s}{Summary statistic vector}
}
\value{
List with mean and sd
}
\description{
Calculate colmeans and colstds
}
