% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scaler}
\alias{scaler}
\title{Scale matrix or vector}
\usage{
scaler(s, mean_and_sd)
}
\arguments{
\item{s}{Matrix or vector}

\item{mean_and_sd}{Mean and sd list (from mean_and_sd)}
}
\value{
scaled matrix or vector
}
\description{
Scale matrix or vector
}
